feat: Move `nsb format` functionality to a class method like `nsb assign` (#20)
Both the RawQuestions and EditedQuestions classes inherit from a base class that defines cached properties that help find various pieces of information. `nsb make` and `nsb format` are now methods of RawQuestions. I had to adjust the tests to account for this new interface, but as a side effect `nsb format` gets the stat reporting of `nsb assign`.

Closes `nsb format` should display the same stats as `nsb assign`  #19